Renamed the ReleaseVersion functions
Renamed Add-ReleaseVersionFolder to Add-FullVersionFolder, and Set-ReleaseVersionPath to Set-FullVersionPath. Updated the comment help for both functions as well as the call to the Set function from within Set-ReleaseFolderStructure.

The purpose of this renaming was to bring those two functions more in line with the other functions that deal with the major version and minor version. This also avoids confusion around the use of release or release version.