Updated some comments and copied more commented-out code
Updated 2 comments in the Main Script section to make the ensuing lines of code more clear.

Also, copied the Start-Release function before the changes to the Commented-out Copied Code section, but only the function name and function comment help.

The plan is to create a Start-Release function that will go through the 4 major tasks like with the existing automation scripts.