Recreated a Start-Release function and added 3 other functions
Recreated the Start-Release function and updated the comment help so the function carries out the 4 major tasks relating to each existing automation script:

- create the release folders
- generate the Python exe
- copy the release content
- generate the release package

Also created the functions Invoke-PythonTool, Copy-ReleaseContent, and New-ReleasePackage for the latter 3 tasks.

Finally, filled out the Start-Release function so it calls Set-ReleaseFolderStructure and the other 3 new functions, in order.

The functions were all named using the approved verbs list for PowerShell commands.