Worked some more on the Invoke-PythonTool function
Tried to work on the section that clears the py-tools folder if it exists and there's stuff in it.

Added some (now commented out) lines to the start of the function to check the difference between what Get-ChildItem returns with and without the Recurse parameter.

Changed the do loop to a foreach loop and changed the $PyInstallerPath variable to the existing py-tool folder for the 1.1-beta release.

However, when testing out the new loop, accidentally deleted the contents of that folder.  So, the pyinstaller command was run again after copying over the necessary files to the py-tool folder.

The foreach loop and associated if-blocks were then adjusted to comment out the actual Remove-Item lines. Rather, the details of the loop were outputted to a file called IPT_test.txt in the project root.

This file shows that the foreach loop doesn't fully iterate through everything. It will end up deleting all specific files, but no folders.