Commented out the foreach loop
Realized that all the work from the past several commits was probably not necessary as the "Remove-Item * -Recurse -Force" line is sufficient to remove everything within the py-tool folder.

All that is needed is to set the location to within the py-tool folder, which is already done. For now, though, the foreach loop code is commented out.