Worked on Set-ReleaseNotesFolder and Set-CompStartFolder
Renamed both functions to use Add instead of Set and updated the comment help to reflect that the functions primarily add the CompStart and release-notes folders if they don't exist.

Updated the code within both functions to use the script-scope variables and to have output that's consistent with the other Add functions.