Rearranged some functions and renamed two functions
Renamed the Confirm-ReleaseFolder to Set-ReleaseFolder and updated the comment help to reflect that the function primarily sets the current location to the release folder.

Renamed the Set-PyToolFolder function to Add-PyToolFolder and updated the comment help to reflect that the function primarily creates a py-tool folder.

Also, started changing the ordering of the functions so all the functions starting Set are at the top and grouped together.