Worked on the Copy-ReleaseContents function
Updated the code to populate the CompStart folder content (that is, the CompStart.bat file, CompStart.ps1 file and config folder).

Also updated the code to copy to the release folder from the assets folder, the release-notes.md and instructions.txt files.

Finally, added some code to copy over the install.ps1 file from the assets folder to the CompStart folder.

However, will need to work on the code more as the directory structure for the release folder CompStart folder will need to be the following in order to accommodate the install.ps1 installer file:

/ - <release-folder>
| - CompStart
  | - install.ps1
  | - installer-files
    | - CompStart
      | - CompStart.ps1
      | - CompStart.bat
      | - config
| - py-tool
| - release-notes
| - instructions.txt