Renamed and refactored Start-Release
Renamed the Start-Release function to Start-ReleaseProcess.

Created an outer loop that will keep presenting the menu to the user until they specifically choose to quit. This way, if the user needs to perform one of the tasks that's not the full release process, they can do so and go back to the menu to perform another task without needing to run the script again.

Also set the Get-ReleaseDetails and Update-PathVars calls to within Start-ReleaseProcess. The user menu includes and option to change the release details. Additionally, the first thing that's checked for, after the user enters their choice, is whether there are any release details. If not, those two functions are called. They are also called if the user chose to change the release details.