Confirmed the Copy-ReleaseContents folder now copies the first 6 files
Fixed some typos in the variable names and string literals which were causing Copy-Item to fail.

The Copy-ReleaseContents function now does the following:

- Set up the CompStart folder and directory structure
- Set up the release-notes folder
- Copy over everything that needs to be in the CompStart folder and its subfolders
- Copy over the release-notes.md file to the release-notes folder

While testing, added an Exit partway through the function so it only does the above and the script quits.

Need to test the code that is supposed to deal with copying over the Python executable.