Finished the Copy-ReleaseContents function
Added 2 new properties to PathVars for the py-tool/dist folder and the actual CompStart.ext paths. There were already analogous properties in FolderNames and FileNames.

Once they were added and set correctly in Update-PathVars, the code to copy the CompStart.exe file was adjusted to use the variables in this script.

The output of the calls to Set-ProjectRoot from within Start-Release was changed to send to $null so the returned result isn't printed to the screen.

A few minor comment and user output changes were also made.