Tested both the DeployRelease and install scripts
After testing the DeployRelease script to create a release called 2.1-alpha (which is not a real release at this time), tested the process of extraction and installation of CompStart.

During that test, discovered that the install.ps1 script created some blank folders in the installed CompStart folder.

Created a folder called testing in /prodenv/scripts and also created issue #82 on GH to deal with this.

Finally, moved all files and folders from both sets of tests to this testing folder.