Cleaned up all testing data
Created a test-data folder in /devenv/data. Moved to it the folder python-tool-testing from /devenv/old-data.

Moved /prodenv/scripts/pre-issue80 to old-data.

Created a folder called deploy-release-script in test-data and moved the two text files in /prodenv/scripts to it.

Finally, moved /prodenv/scripts/testing to deploy-release-script and renamed it to test_2025-01-30.