Updated 3 READMEs and created 2 READMEs
Updated the project root README.

Started updating the packages README and the releases README.

Also started a README.md file for assets and scripts.

There is old information in some of hte README files that is marked as such. This is so the information can be modified or deleted as needed.