Added some old README information
Prior to the merging of issue #81 and issue #84, while resolving the conflicts found, some potentially useful description data was copied into a local text file. This was done because, even though the resolved files didn't contain the same data, other README files might be able to use the data.

That copy-pasted data was added to a text file in the project root to work on as appropriate or needed.