After testing, discovered same issue
The previous change didn't fix the issue. The annotations were updated to reflect this and discuss what could be done.

Copy-pasting one annotation that explains the actual problem:

It looks like the problem is that, instead of using a simple Copy-Item -Recurse, a list of the files and folders in $InstallFullPath is taken but then copied over one by one in a loop. The list is populated using Get-ChildItem -Recurse, which correctly preserves the directory structure. But then each item is copied to the CompStart folder in %localappdata%. So, the directory structure isn't recreated.