Remade the change to $FilesList in install.ps1
Realized that the change made in commit 7445fcd was made to the wrong install.ps1 file.

The commit details for that commit are:

> Refactored how the $FilesList is created
> Set up the $InstallFullPath variable by using $PSScriptRoot. This should ensure the proper parent folder is always correct. The variable $InstallerFolder is then added to it.

While it was later discovered in commit 63b0610 that this changed didn't fix issue #82, the change was made to the install.ps1 script in /devenv/data/test-data/release-install-script/test_2025-01-30/extraction/CompStart.

Made the same change to the correct install.ps1 script, found at /prodenv/assets/cs-installer. From now on, will need to make any changes to this correct installer script.

Since the process to deploy a release works and can be used for testing purposes, and since there's the /devenv/test-data/release-install-script folder to hold any and all data generated while testing possible fixes, working only with the prodenv version of the installer script makes the most sense.