Created a new test release for the installer script
The new test is in /devenv/data/test-data/release-install-script/test_2025-02-12. The contents of the test is the exact same as for test_2025-01-30 because nothing was changed.

However, this second test was run to be able to step through the install.ps1 script. This allowed for a refreshing of the state of things and also gave some ideas on a potential fix.