Separated the installation into two sections
The first section is to create the directory structure for CompStart and the second one is to copy over all the files.

This allows for the subfolders to be created using the correct tree structure.

In the files installation section, the parent path of each file item is compared with the destination path and updated as needed to ensure the correct file is installed to the correct subfolder.