Finished updating DirectoryStructure.md
Went through all of the folders listed and updated them. Some were removed as they are no longer relevant, and some just had their descriptions refactored.

Will need to go through this file with fresh eyes to see if the L4 folders are needed or not. It may be possible to only keep the ones for /devenv/config/schema and /devenv/data/test-data.