Created the fake release 2.1-main to update the /docs/prodenv-docs/RELEASES.md file, but while reviewing the release and package files that were created, realized a few things
While looking at the /prodenv/packages/v2/m1/CompStart-2.1-main.zip file, noted that the packaging process will need to be modified. This is because the archive structure is the following:

CompStart \
|
| - install.ps1
| - installer-files \
| - |
| - | - instructions.txt
| - | - CompStart \
| - | - |
| - | - | - CompStart.bat
| - | - | - CompStart.exe
| - | - | - CompStart.ps1
| - | - | - config \

In other words, the instructions.txt file, which is supposed to have installation instructions, is currently one level down from the install.ps1 file. Additionally, there's an install.bat file found in /prodenv/assets/cs-installer/ that should be included as well in the package, since this is meant to facilitate easy running of the PowerShell script.

Will need to think about how to move forward, but may need to close issue #81 for now, open a new one to update the installer script, and then resume updating the 5 documentation files that are left.