Updated PACKAGES.md
Rewrote the Packaging section to explain both the ZIP file contents and how to create a package.

For how to create a package, used the same paragraph as how to create a release in RELEASES.md. In both cases, the paragraphs specifies to run the DeployRelease.ps1 script and references SCRIPTS.md for more information.